requirements.txt
flask
flask-cors
torch
torchvision
numpy
opencv-python
pillow
matplotlib
pyspellchecker==0.7.1
